+++ tests/result_debug_line.c
@@ -21,16 +21,19 @@
  * Tests `RESULT_DEBUG_LINE`.
  */
 int main() {
-#ifdef NDEBUG
-    TEST_SKIP("NDEBUG is defined");
-#else
     // Given
     RESULT_STRUCT(int, char);
     const RESULT(int, char) success = RESULT_SUCCESS(512);
     const RESULT(int, char) failure = RESULT_FAILURE('A');
     // Then
-    TEST_ASSERT_INT_EQUALS(RESULT_DEBUG_LINE(success), 29);
-    TEST_ASSERT_INT_EQUALS(RESULT_DEBUG_LINE(failure), 30);
-    TEST_PASS;
+#ifdef NDEBUG
+    (void) success;
+    (void) failure;
+    TEST_ASSERT_INT_EQUALS(RESULT_DEBUG_LINE(success), 0);
+    TEST_ASSERT_INT_EQUALS(RESULT_DEBUG_LINE(failure), 0);
+#else
+    TEST_ASSERT_INT_EQUALS(RESULT_DEBUG_LINE(success), 26);
+    TEST_ASSERT_INT_EQUALS(RESULT_DEBUG_LINE(failure), 27);
 #endif
+    TEST_PASS;
 }